Move benchar to libraries
