+++ .gitignore
@@ -1,2 +1,2 @@
 backup*/*
-benchar/build/*
+lib/benchar/build/*
