+++ lib/benchar/cbenchar.cpp
