+++ test/test_benchar.py
@@ -2,10 +2,10 @@ import unittest
 import os
 
 from exact_string_matching import forward
-from benchar import benchar
+from lib.benchar import benchar
 
 if os.environ.get('CBENCHAR') is not None:
-    from benchar.build import cbenchar
+    from lib.benchar.build import cbenchar
 
 class TestBenchar(unittest.TestCase):
     run_cbenchar = unittest.skipUnless(
